AI_tool/requirements.txt
streamlit
google-generativeai
pdfplumber
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv